Show and switch to a window instead of a session
The format shown to the user is slightly modified, see here the
difference:

Before:
    dotfiles: 2 window(s) (attached)
    thesis: 1 window(s)
    /home/user/Downloads
    ...

After:
    dotfiles/zsh (attached)
    dotfiles/tmux
    thesis/toolset
    /home/user/Downloads
    ...

This means that RESULT is modified as well. There is no colon in
there, which means that we have to check whether the string begins
with a /. If it does, RESULT is a path. If not, it must be a session.
This has two downsides:

  - It might not work on Windows, since a path probably does not start
  with / on Windows. I do not know how Windows paths work (do they
  start with \\, Or with C:\?), and I cannot test it.

  - If the session_name starts with /, the check fails because we
  assume that it is a path instead. This last part I have let it
  slide, since the current script is also not really forgiving if the
  session_name contains a space (and another session_name exists with
  the first word as beginning as well).